Change Railway Tracks to Physics
Update the name to include ``(Equilibria)'' to make the main scientific topic clearer.

Changed subject from ``Engineering'' to ``Physics''. There are arguments that this could be an engineering demo if you focus on designing train wheels however the current setup means (in my opinion) the main focus is on more physics-related topics of equilibria, system dynamics, and stable and unstable equilibria.