Update Status of Exploring Genes
Only the DNA Model stand is currently broken. The model can easily be used without the stand.

The Gilson Pipette is currently missing (or maybe we returned it and forgot...). From a quick look, they are available on ebay for £40-60 second hand, although I'm unsure if they are the correct type and not convinced how much it adds to the experiment since it's used to demonstrate how small an amount of liquid would be used and then running a fake DNA gel... Maybe upto a biologist to determine if we need to aquire one and if so which one.